pyproject.toml
[tool.poetry]
name = "proxai"
version = "0.1.0"
description = "ProxAI is a lightweight abstraction layer for foundational AI model connections."
authors = ["proxai"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
openai = "^1.14.3"
anthropic = "^0.21.3"
google-generativeai = "^0.4.1"
cohere = "^5.1.7"
databricks-genai-inference = "^0.2.3"
mistralai = "^0.1.8"

[tool.poetry.dev-dependencies]

[tool.poetry.group.dev.dependencies]
pytest = "^8.1.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
